+++ projects/packages/backup/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.2.26] - 2025-10-20
+### Changed
+- Update dependencies. [#45488]
+
 ## [4.2.25] - 2025-10-06
 ### Changed
 - Update package dependencies. [#45334] [#45335]
@@ -959,6 +963,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add API endpoints and Jetpack Backup package for managing Help…
 
+[4.2.26]: https://github.com/Automattic/jetpack-backup/compare/v4.2.25...v4.2.26
 [4.2.25]: https://github.com/Automattic/jetpack-backup/compare/v4.2.24...v4.2.25
 [4.2.24]: https://github.com/Automattic/jetpack-backup/compare/v4.2.23...v4.2.24
 [4.2.23]: https://github.com/Automattic/jetpack-backup/compare/v4.2.22...v4.2.23