+++ projects/packages/external-media/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.7] - 2025-10-20
+### Changed
+- External media: Prevent site editor performance issues by bringing back previous script enqueue approach. [#45546]
+- Update dependencies. [#45488]
+
 ## [0.5.6] - 2025-10-06
 ### Changed
 - Update package dependencies. [#45334]
@@ -198,6 +203,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix the button size in the editor for Gutenberg 18 or below. [#41619]
 - Media Library: Fix the Import Media button color in some color schemes. [#41664]
 
+[0.5.7]: https://github.com/Automattic/jetpack-external-media/compare/v0.5.6...v0.5.7
 [0.5.6]: https://github.com/Automattic/jetpack-external-media/compare/v0.5.5...v0.5.6
 [0.5.5]: https://github.com/Automattic/jetpack-external-media/compare/v0.5.4...v0.5.5
 [0.5.4]: https://github.com/Automattic/jetpack-external-media/compare/v0.5.3...v0.5.4